Add 'data-click' attribute to menu to prevent duplicate on click handlers
The cause for the bug is that when user navigates to their profile via
navbar-menu the page is not refreshed and as a result another click event
listener is added to the existing one. Thus the navbar's class list gets
toggled twice, which in the end removes 'showing' class.

The fix is to add 'data-click' attribute to the navigation button to signify
that a click listener is already present on it, in which case a duplicate one
will not be added.